Filter during DP for duplicate units (of which there are a ton)
Duplicate units seem to grow, although the core cause wasn't known.
It was exactly powers of two, which seem to suggest that it's because
of something to do with the parity of the reciprocal, but I'm not sure
how it could explode like that. Maybe strings with a lot of repetition?
After filtering for repeats, the DP is lightning-fast.